+++ lib/src/api/api_client_stub.dart
@@ -1,4 +1,0 @@
-import 'package:dio/dio.dart';
-
-/// No-op على الويب - المتصفح يتولى SSL.
-void configureAllowBadCertificate(Dio dio, {bool allow = true}) {}